Update the license for StereoCancer to MIT from GPL.
The original reason for licensing it under GPL was because I thought the IsMirrorCheck code was original to CancerSpace. The original source for this code is a post by Merlin on 7/20/2018 at 1:21 AM in the VRC Shader Dev discord. This was confirmed by talking with AkaiMage a long time ago.
Additionally, the IsMirrorCheck code doesn't even exist in the codebase anymore as I utilize the VRChat global uniforms for checking cameras and mirrors now. The shader itself, StereoCancer, was also entirely written without looking at CancerSpace. Turning on the Bold FontStyle was also from CancerSpace, but that has now been deleted and replaced with a thick monospace Google Font (Oswald).